srcs/app/requirements.txt
Django==5.1
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
whitenoise==6.7.0
python-dotenv==1.0.1
pillow==10.4.0
asgiref==3.8.1
sqlparse==0.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
constantly==23.10.4
cryptography==43.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.1.1
service-identity==24.1.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.1.0
web3==7.4.0
bleach==6.2.0